Modulerized all code to follow SOLID standards
